test(scroll-gradient): use separate mock
